Make scoring work for both players
The boundary on the bottom of the screen was never triggering a point
because the collision detection within the "Ball" class checks if the
ball is going to pass the screen boundary based on its size so the
actual "Position" of the ball would never pass the boundary and so never
score a point.

Indicate that the ball collided with a boundary on the Y axis so that
the game scoring system can detect the boundary collision and assign a
point to the appropriate player.